pyproject.toml
[tool.poetry]
name = "PyDo"
version = "0.0.1"
description = "A Todo List API made with Python, FastAPI, SQLAlchemy and a lot of work."
authors = ["Vinícius Oliveira <eu@viniciusoliveira.me>"]

[tool.poetry.dependencies]
python = "^3.8"
aiomysql = "0.0.21"
alembic = "1.6.5"
appdirs = "1.4.4"
asgiref = "3.4.0"
CacheControl = "0.12.6"
cachy = "0.3.0"
certifi = "2021.5.30"
chardet = "4.0.0"
cleo = "0.8.1"
click = "7.1.2"
clikit = "0.6.2"
colorama = "0.4.4"
commonmark = "0.9.1"
crashtest = "0.3.1"
databases = "0.4.3"
distlib = "0.3.2"
fastapi = "0.65.2"
FastAPI-SQLAlchemy = "0.2.1"
filelock = "3.0.12"
greenlet = "1.1.0"
h11 = "0.12.0"
html5lib = "1.1"
idna = "2.10"
keyring = "21.8.0"
lockfile = "0.12.2"
Mako = "1.1.4"
MarkupSafe = "2.0.1"
msgpack = "1.0.2"
mysqlclient = "2.0.3"
packaging = "20.9"
pastel = "0.2.1"
pexpect = "4.8.0"
pipenv-poetry-migrate = "0.1.5"
pkginfo = "1.7.0"
poetry = "1.1.7"
poetry-core = "1.0.3"
ptyprocess = "0.7.0"
pydantic = "1.8.2"
Pygments = "2.9.0"
pylev = "1.4.0"
PyMySQL = "0.9.3"
pyparsing = "2.4.7"
python-dateutil = "2.8.1"
python-dotenv = "0.18.0"
python-editor = "1.0.4"
requests = "2.25.1"
requests-toolbelt = "0.9.1"
rich = "9.13.0"
shellingham = "1.4.0"
six = "1.16.0"
SQLAlchemy = "1.3.24"
starlette = "0.14.2"
tomlkit = "0.7.2"
typer = "0.3.2"
typing-extensions = "3.10.0.0"
urllib3 = "1.26.6"
uvicorn = "0.14.0"
virtualenv = "20.4.7"
webencodings = "0.5.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
